Commander: start pulling arming related parts into separate folder
