+++ src/lib/systemlib/mavlink_log.h
@@ -41,6 +41,7 @@
 
 #pragma once
 
+#include <px4_log.h>
 #include <uORB/uORB.h>
 
 /**
