+++ src/modules/commander/Arming/PreFlightCheck/CMakeLists.txt
@@ -1,6 +1,6 @@
 ############################################################################
 #
-#   Copyright (c) 2015 PX4 Development Team. All rights reserved.
+#   Copyright (c) 2019 PX4 Development Team. All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
 # modification, are permitted provided that the following conditions
@@ -31,39 +31,8 @@
 #
 ############################################################################
 
-add_subdirectory(failure_detector)
-add_subdirectory(Arming)
-
-px4_add_module(
-	MODULE modules__commander
-	MAIN commander
-	COMPILE_FLAGS
-	SRCS
-		accelerometer_calibration.cpp
-		airspeed_calibration.cpp
-		baro_calibration.cpp
-		calibration_routines.cpp
-		Commander.cpp
-		commander_helper.cpp
-		esc_calibration.cpp
-		gyro_calibration.cpp
-		health_flag_helper.cpp
-		mag_calibration.cpp
-		PreflightCheck.cpp
-		rc_calibration.cpp
-		rc_check.cpp
-		state_machine_helper.cpp
-	DEPENDS
-		circuit_breaker
-		df_driver_framework
-		failure_detector
-		git_ecl
-		ecl_geo
-		hysteresis
-		PreFlightCheck
-		ArmAuthorization
-	)
-
-if(PX4_TESTING)
-	add_subdirectory(commander_tests)
-endif()
+px4_add_library(PreFlightCheck
+	PreFlightCheck.cpp
+)
+target_include_directories(PreFlightCheck PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
+target_link_libraries(PreFlightCheck PUBLIC ArmAuthorization)
